"Replace ContractPlan with CampaignPriority in campaign logic"
Replaced ContractPlan with a newly introduced CampaignPriority in the methods related to campaign logic in WaitingListLogic, CampaignRewardRequest, CampaignCode, ContractPlan, CampaignEvents and CampaignRule classes. The change allows for a more encompassing code structure, enabling priority to be taken into account in campaign events, which was not reflective in the previous 'plan' attribute. This offers a more accurate representation of a customer's campaign priority over mere contractual plans, facilitating better classification in campaign events and reward systems.