"Renamed WaitingListRuleId to CampaignRuleId"
Refactoring of WaitingListRuleId to CampaignRuleId is completed across different classes including WaitingListLogic, CampaignEvents, and CampaignRule. This name change was important to avoid confusion because this ID is used for not only the waiting list, but also other campaign rules. This new naming convention will also support future extension of the campaign feature as it provides a more accurate and generalized representation.