Simplify addNewWaitingList parameters
Refactored the addNewWaitingList method to simplify the parameters. Previously, the method was passing large number of parameters which made it hard to manage and prone to errors. Now, dependencies on objects like CampaignEvents, IdGenerator, URIBuilder, WaitingList etc. are handled internally within the AddNewWaitingList class itself, thereby reducing the number of method parameters and enhancing the code readability.