Change save methods to 'register' in WaitingListLogic
To clarify the purpose of the 'save' methods in WaitingListLogic, they have been renamed to 'register'. The classes impacted by this change include SaveNewWaitingList, SaveAsNewBookingWithCampaignReward, AddNewWaitingList, and SaveAsBookingIfAvailable. This improves the intuitiveness of the code, making it easier for newcomers to understand it. It also aligns better with the terminologies used in the particular use-cases.