Refactor getStrategy method in WaitingListLogic class
The getStrategy method in the WaitingListLogic class has been updated to introduce the CampaignApplicationStrategyFactory object as a method parameter. This change has been implemented to improve code readability and fluency by avoiding repetitive strategy calculations for products that do not have a waiting list. Additionally, this introduces a more streamlined workflow, reducing possibilities of code repetitiveness and inconsistencies in the behavior of the system.