Simplify WaitingListLogic by leveraging StrategyFactory
Refactored WaitingListLogic to rely on CampaignApplicationStrategyFactory instead of spawning instances of several distinct classes. This change results in a leaner code structure, improves maintainability, and adheres to the Dependency Inversion Principle. This change calls the getStrategy method directly on the CampaignApplicationStrategyFactory instance, thus eliminating the need for the previously used getStrategy method in WaitingListLogic.